requirements.txt
Flask>=2.2
python-dotenv>=0.21.0
gunicorn>=20.0.0
# Dev / test packages (optional)
pytest>=7.0
pytest-flask>=1.2
